Move still image callback to MapContext
